move augmentWallet to provider
