+++ packages/govern-console/src/components/ActionBuilder/Screens/Deposit.tsx
@@ -17,7 +17,7 @@ import {
 import { Hint } from 'components/Hint/Hint';
 import { useForm, Controller } from 'react-hook-form';
 import { validateAmountForDecimals, validateToken, validateBalance } from 'utils/validations';
-import { useWallet } from 'AugmentedWallet';
+import { useWallet } from 'providers/AugmentedWallet';
 import { useActionBuilderState } from '../ActionBuilderStateProvider';
 import { getTruncatedAccountAddress } from 'utils/account';
 import { getErrorFromException } from 'utils/HelperFunctions';
