+++ packages/govern-console/src/hooks/proposal-hooks.tsx
@@ -1,4 +1,4 @@
-import { useWallet } from 'AugmentedWallet';
+import { useWallet } from 'providers/AugmentedWallet';
 import { useMemo } from 'react';
 import QueueApprovals from 'services/QueueApprovals';
 import FacadeProposal from 'services/Proposal';
